# perf(bash-v2): speed up filtering menu-complete descriptions