Similarly as fb8031162c2ffab270774f13c6904bb04cbba5a7 (+ the empty entry fix in https://github.com/spf13/cobra/pull/1691) did for the "regular", non-menu completion cases.

I have no numbers on this, but I'm assuming the speedups seen elsewhere are here too, and didn't notice anything breaking in casual manual tests.

Could be useful to add some perf numbers for the menu-complete code path too to [marckhouzam/cobra-completion-testing](https://github.com/marckhouzam/cobra-completion-testing/) (maybe as well as other tests, haven't checked if there are any).