@marckhouzam 2022-05-04
> Could be useful to add some perf numbers for the menu-complete code path too 

Here's a PR that adds performance tests for menu-complete: https://github.com/marckhouzam/cobra-completion-testing/pull/16

It confirms that your current optimisation gives a 10x improvement!
Before: 
```
menu-complete no descs: 1000 completions took 1.051 seconds
menu-complete with descs: 1000 completions took 1.062 seconds
```
With the current PR:
```
menu-complete no descs: 1000 completions took 0.103 seconds
menu-complete with descs: 1000 completions took 0.128 seconds
```